# apps/api/src/swagger/swagger_ui.json
{
	"swagger": "2.0",
	"info": {
		"version": "1.0.0",
		"title": "POP Api",
		"description": "Documentation sur les webservices de l'application Api"
	},
	"host": "localhost:3000",
	"basePath": "/",
...
	"schemes": ["http"],
	"securityDefinitions": {
		"apiKeyAuth": {
			"type": "apiKey",
			"in": "header",
			"name": "X-API-KEY",
			"description": "any description..."
		}
	},
	"consumes": ["application/json"],
	"produces": ["application/json"],
	"paths": {
...
		"/thesaurus/search": {
			"get": {
				"tags": ["Thesaurus"],
				"description": "Retourne les informations de la notice Thesaurus en fonction des paramètres",
				"parameters": [
					{
						"name": "id",
						"in": "query",
						"type": "string"
					},
					{
						"name": "value",
						"in": "query",
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					}
				}
			}
		},
		"/thesaurus/autocompleteThesaurus": {
			"get": {
				"tags": ["Thesaurus"],
				"description": "Retourne les thésaurus en fonction de son identifiant et correspondant à la valeur (autocompletion)",
				"parameters": [
					{
						"name": "id",
						"in": "query",
						"type": "string"
					},
					{
						"name": "value",
						"in": "query",
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					}
				}
			}
		},
		"/thesaurus/validate": {
			"get": {
				"tags": ["Thesaurus"],
				"description": "Retourne les thésaurus en fonction de son identifiant et correspondant à la valeur",
				"parameters": [
					{
						"name": "id",
						"in": "query",
						"type": "string"
					},
					{
						"name": "value",
						"in": "query",
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					}
				}
			}
		},
		"/thesaurus/getTopConceptsByThesaurusId": {
			"get": {
				"tags": ["Thesaurus"],
				"description": "Retourne les informations de la notice Thesaurus en fonction des paramètres",
				"parameters": [
					{
						"name": "id",
						"in": "query",
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"500": {
						"description": "Internal Server Error"
					}
				}
			}
		},
		"/thesaurus/getAllChildrenConcept": {
			"get": {
				"tags": ["Thesaurus"],
				"description": "Retourne les informations de la notice Thesaurus en fonction des paramètres",
				"parameters": [
					{
						"name": "id",
						"in": "query",
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"500": {
						"description": "Internal Server Error"
					}
				}
			}
		},
		"/thesaurus/getPreferredTermByConceptId": {
			"get": {
				"tags": ["Thesaurus"],
				"description": "Retourne les informations de la notice Thesaurus en fonction des paramètres",
				"parameters": [
					{
						"name": "id",
						"in": "query",
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"500": {
						"description": "Internal Server Error"
					}
				}
			}
		},
		"/thesaurus/deleteAllThesaurus": {
			"get": {
				"tags": ["Thesaurus"],
				"description": "Suppression de tous les Thesaurus",
				"parameters": [
					{
						"name": "id",
						"in": "query",
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"500": {
						"description": "Internal Server Error"
					}
				}
			},
			"post": {
				"tags": ["Thesaurus"],
				"description": "Création d",
				"parameters": [
					{
						"name": "id",
						"in": "query",
						"type": "string"
					},
					{
						"name": "body",
						"in": "body",
						"schema": {
							"type": "object",
							"properties": {
								"terms": {
									"example": "any"
								}
							}
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"500": {
						"description": "Internal Server Error"
					}
				}
			}
		},
		"/thesaurus/getThesaurusById": {
			"get": {
				"tags": ["Thesaurus"],
				"description": "Retourne la liste des Thesaurus en fonction de l",
				"parameters": [
					{
						"name": "id",
						"in": "query",
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"400": {
						"description": "Bad Request"
					},
					"401": {
						"description": "Unauthorized"
					}
				}
			}
		},
...
		"/thesaurus/autocompleteByIdthesaurusAndValue": {
			"get": {
				"tags": ["Thesaurus"],
				"description": "Recherche de l",
				"parameters": [
					{
						"name": "id",
						"in": "query",
						"type": "string"
					},
					{
						"name": "value",
						"in": "query",
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"400": {
						"description": "Bad Request"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/thesaurus/getAllThesaurusById": {
			"get": {
				"tags": ["Thesaurus"],
				"description": "Retourne la liste des Thesaurus en fonction de l'identifiant",
				"parameters": [
					{
						"name": "id",
						"in": "query",
						"type": "string"
					}
				],
				"responses": {
					"202": {
						"description": "Accepted"
					}
				}
			}
		},
		"/thesaurus/getPrefLabelByIdArk": {
			"get": {
				"tags": ["Thesaurus"],
				"description": "Recherche le prefLabel par rapport à l",
				"parameters": [
					{
						"name": "id",
						"in": "query",
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"400": {
						"description": "Bad Request"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
...
	},
...
}